Created separate unit test files for Vitals and Diagnosis.
